+++ src/variable.js
@@ -1,7 +1,10 @@
+import formatInputValue from './format-input-value';
+
 export class VariableDefinition {
-  constructor(name, type) {
+  constructor(name, type, defaultValue) {
     this.name = name;
     this.type = type;
+    this.defaultValue = defaultValue;
     Object.freeze(this);
   }
 
@@ -10,7 +13,9 @@ export class VariableDefinition {
   }
 
   toVariableDefinitionString() {
-    return `$${this.name}:${this.type}`;
+    const defaultValueString = this.defaultValue ? ` = ${formatInputValue(this.defaultValue)}` : '';
+
+    return `$${this.name}:${this.type}${defaultValueString}`;
   }
 }
 
@@ -18,6 +23,6 @@ export function isVariable(value) {
   return value instanceof VariableDefinition;
 }
 
-export default function variable(name, type) {
-  return new VariableDefinition(name, type);
+export default function variable(name, type, defaultValue) {
+  return new VariableDefinition(name, type, defaultValue);
 }
